Support SECP256r1 and SECP521r1 curves (#102)
* Tar and workflow fixes and tests


* Create locks even if capsule is not supported

* Update cdoc/CDoc2Reader.cpp


* Use toUint8Vector helper

* Removed some extra check from parsing

* Remove INVALID lock type and rename UNSUPPORTED to UNKNOWN

* Return NULL from frontend if CDoc constructor fails

* Remove label creation from Java test

* Keep Java references in CDocReader and CDocWriter to prevent premature GC

* Cleaned up logging

* Renamed ILogger -> Logger

* Added Logger.h

* Removed ConsoleLogger from public interface

* Fixed WinBackend.cpp

* Fixed WinBackend.cpp more

* One more Winbackend fix

* Implement secp256r1 curve support

* Added CDoc.h

* Added cdoc-tool support and testing for secp256 curve

* Allow unknown curve type while decrypting

* Minor cleanups

* Removed Algorithm enum from Certificate and PKCS11 cleanup

* Export CDocCipher internally for cdoc-tool

* Fix label parsing

* Handle "PaxHeaders" paths correctly

* Apply suggestions from code review


* Fix tar filenames if only PAX size is used

* Moved parseLabel to Lock, take minimum of expiry time

* Fixed CDocTool.java

* Use property system for automatic label generation

* remove final from CDoc2

* Exclude CDoc2 from swig for the time being

* Remove commented-out code

* Remove 'final' once more

* Made CDoc2 back to namespace

* Moved label strings to Label namespace

* #undef VERSION

* Check recipient validity in addRecipient

* Apply suggestions from code review


* Do not truncate user-specified expiry to 6 months, force cdoc-tool to utf8 codepage

* Comment out keyshares support

* Set version to 0.5.0

* Remove forced Debug build

* Fixed Recipient

* Added secp521r1 curve support

* Remove unused variable

* Added secp521r1 test keys/certificates

* Fix recipient

* Cleaned Recipient building

* Unify server and public key recipient constructors

---------

+++ test/libcdoc_boost.cpp
@@ -51,9 +51,15 @@ constexpr string_view SourceFile3("test_data3.txt");
  * @brief Encrypted file name.
  */
 constexpr string_view TargetFile("test_data.txt.cdoc");
-constexpr string_view ECPrivKeyFile("ec-secp384r1-priv.der");
-constexpr string_view ECPubKeyFile("ec-secp384r1-pub.der");
-constexpr string_view ECCertFile("ec-secp384r1-cert.der");
+constexpr string_view EC384PrivKeyFile("ec-secp384r1-priv.der");
+constexpr string_view EC384PubKeyFile("ec-secp384r1-pub.der");
+constexpr string_view EC384CertFile("ec-secp384r1-cert.der");
+constexpr string_view EC256PrivKeyFile("ec-secp256r1-priv.der");
+constexpr string_view EC256PubKeyFile("ec-secp256r1-pub.der");
+constexpr string_view EC256CertFile("ec-secp256r1-cert.der");
+constexpr string_view EC521PrivKeyFile("ec-secp521r1-priv.der");
+constexpr string_view EC521PubKeyFile("ec-secp521r1-pub.der");
+constexpr string_view EC521CertFile("ec-secp521r1-cert.der");
 constexpr string_view RSAPrivKeyFile("rsa_2048_priv.der");
 constexpr string_view RSAPubKeyFile("rsa_2048_pub.der");
 constexpr string_view RSACertFile("rsa_2048_cert.der");
@@ -277,7 +283,7 @@ encryptV2(const std::vector<std::string>& files, const std::string& container, c
 }
 
 static void
-decrypt(const std::vector<std::string>& files, const std::string& container, const std::string& dir, libcdoc::RcptInfo& rcpt)
+decrypt(const std::vector<std::string>& files, const std::string& container, const std::string& dir, libcdoc::RcptInfo& rcpt, bool remove = true)
 {
     libcdoc::ToolConf conf;
     conf.input_files.push_back(container);
@@ -292,7 +298,7 @@ decrypt(const std::vector<std::string>& files, const std::string& container, con
     }
 
     path = fs::path(container);
-    if (fs::exists(path)) {
+    if (remove && fs::exists(path)) {
         error_code e;
         fs::remove(path, e);
         if(e)
@@ -301,10 +307,10 @@ decrypt(const std::vector<std::string>& files, const std::string& container, con
 }
 
 static void
-decrypt(const std::vector<std::string>& files, const std::string& container, const std::string& dir, const std::vector<uint8_t>& key)
+decrypt(const std::vector<std::string>& files, const std::string& container, const std::string& dir, const std::vector<uint8_t>& key, int idx = 0, bool remove = true)
 {
-    libcdoc::RcptInfo rcpt {.type=libcdoc::RcptInfo::LOCK, .secret=key, .lock_idx=0};
-    decrypt(files, container, dir, rcpt);
+    libcdoc::RcptInfo rcpt {.type=libcdoc::RcptInfo::LOCK, .secret=key, .lock_idx=idx};
+    decrypt(files, container, dir, rcpt, remove);
 }
 static int
 unicode_to_utf8 (unsigned int uval, uint8_t *d, uint64_t size)
@@ -559,60 +565,38 @@ BOOST_FIXTURE_TEST_CASE_WITH_DECOR(DecryptWithPasswordLabelIndex, DecryptFixture
 }
 BOOST_AUTO_TEST_SUITE_END()
 
-// CDoc2 AES key
+// CDoc2 public/private/symmetric key
 
-BOOST_AUTO_TEST_SUITE(AESKeyUsage)
-BOOST_FIXTURE_TEST_CASE_WITH_DECOR(EncryptWithAESKey, EncryptFixture,
-        * utf::description("Encrypting a file with symmetric AES key"))
+BOOST_AUTO_TEST_SUITE(CDoc2KeyUsage)
+static constexpr string_view CONTAINER("CDoc2KeyUsage.cdoc");
+BOOST_FIXTURE_TEST_CASE_WITH_DECOR(EncryptWithCDoc2Key, EncryptFixture,
+        * utf::description("Encrypting a CDoc2 file with a key"))
 {
     std::vector<libcdoc::RcptInfo> rcpts {
+        {libcdoc::RcptInfo::PKEY, {}, {}, fetchDataFile(EC384PubKeyFile)},
+        {libcdoc::RcptInfo::CERT, {}, fetchDataFile(EC384CertFile)},
+        {libcdoc::RcptInfo::PKEY, {}, {}, fetchDataFile(EC256PubKeyFile)},
+        {libcdoc::RcptInfo::CERT, {}, fetchDataFile(EC256CertFile)},
+        {libcdoc::RcptInfo::PKEY, {}, {}, fetchDataFile(EC521PubKeyFile)},
+        {libcdoc::RcptInfo::CERT, {}, fetchDataFile(EC521CertFile)},
+        {libcdoc::RcptInfo::PKEY, {}, {}, fetchDataFile(RSAPubKeyFile)},
         {libcdoc::RcptInfo::SKEY, "AES", {}, libcdoc::fromHex(AESKey)}
     };
-    encrypt(2, {checkDataFile(sources[0])}, formTargetFile("AESKeyUsage.cdoc"), rcpts);
+    encrypt(2, {checkDataFile(sources[0])}, formTargetFile(CONTAINER), rcpts);
 }
-BOOST_FIXTURE_TEST_CASE_WITH_DECOR(DecryptWithAESKey, DecryptFixture,
-                     * utf::depends_on("AESKeyUsage/EncryptWithAESKey")
-                     * utf::description("Decrypting a file with with symmetric AES key"))
-{
-    decrypt({checkDataFile(sources[0])}, checkTargetFile("AESKeyUsage.cdoc"), tmpDataPath.string(), libcdoc::fromHex(AESKey));
-}
-BOOST_AUTO_TEST_SUITE_END()
 
-// CDoc2 EC public/private key
-
-BOOST_AUTO_TEST_SUITE(ECKeyUsage)
-BOOST_FIXTURE_TEST_CASE_WITH_DECOR(EncryptWithECKey, EncryptFixture,
-        * utf::description("Encrypting a file with EC key"))
+BOOST_FIXTURE_TEST_CASE_WITH_DECOR(DecryptWithCDoc2Key, DecryptFixture,
+                     * utf::depends_on("CDoc2KeyUsage/EncryptWithCDoc2Key")
+                     * utf::description("Decrypting a CDoc2 file with a key"))
 {
-    std::vector<libcdoc::RcptInfo> rcpts {
-        {libcdoc::RcptInfo::PKEY, {}, {}, fetchDataFile(ECPubKeyFile)}
-    };
-    encrypt(2, {checkDataFile(sources[0])}, formTargetFile("ECKeyUsage.cdoc"), rcpts);
-}
-BOOST_FIXTURE_TEST_CASE_WITH_DECOR(DecryptWithECKey, DecryptFixture,
-                     * utf::depends_on("ECKeyUsage/EncryptWithECKey")
-                     * utf::description("Decrypting a file with with EC private key"))
-{
-    decrypt({checkDataFile(sources[0])}, checkTargetFile("ECKeyUsage.cdoc"), tmpDataPath.string(), fetchDataFile(ECPrivKeyFile));
-}
-BOOST_AUTO_TEST_SUITE_END()
-
-// CDoc2 RSA public/private key
-
-BOOST_AUTO_TEST_SUITE(RSAKeyUsage)
-BOOST_FIXTURE_TEST_CASE_WITH_DECOR(EncryptWithRSAKey, EncryptFixture,
-        * utf::description("Encrypting a file with RSA key"))
-{
-    std::vector<libcdoc::RcptInfo> rcpts {
-        {libcdoc::RcptInfo::PKEY, {}, {}, fetchDataFile(RSAPubKeyFile)}
-    };
-    encrypt(2, {checkDataFile(sources[0])}, formTargetFile("RSAKeyUsage.cdoc"), rcpts);
-}
-BOOST_FIXTURE_TEST_CASE_WITH_DECOR(DecryptWithRSAKey, DecryptFixture,
-                     * utf::depends_on("RSAKeyUsage/EncryptWithRSAKey")
-                     * utf::description("Decrypting a file with with RSA private key"))
-{
-    decrypt({checkDataFile(sources[0])}, checkTargetFile("RSAKeyUsage.cdoc"), tmpDataPath.string(), fetchDataFile(RSAPrivKeyFile));
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), fetchDataFile(EC384PrivKeyFile), 0, false);
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), fetchDataFile(EC384PrivKeyFile), 1, false);
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), fetchDataFile(EC256PrivKeyFile), 2, false);
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), fetchDataFile(EC256PrivKeyFile), 3, false);
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), fetchDataFile(EC521PrivKeyFile), 4, false);
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), fetchDataFile(EC521PrivKeyFile), 5, false);
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), fetchDataFile(RSAPrivKeyFile), 6, false);
+    decrypt({checkDataFile(sources[0])}, checkTargetFile(CONTAINER), tmpDataPath.string(), libcdoc::fromHex(AESKey), 7, true);
 }
 BOOST_AUTO_TEST_SUITE_END()
 
@@ -622,13 +606,13 @@ BOOST_AUTO_TEST_SUITE(CDoc1ECKeySingle)
 BOOST_FIXTURE_TEST_CASE_WITH_DECOR(EncryptWithECKeyV1, EncryptFixture,
         * utf::description("Encrypting a file with EC key in CDoc1 format"))
 {
-    encryptV1({checkDataFile(sources[0])}, formTargetFile("ECKeyUsageV1.cdoc"), fetchDataFile(ECCertFile));
+    encryptV1({checkDataFile(sources[0])}, formTargetFile("ECKeyUsageV1.cdoc"), fetchDataFile(EC384CertFile));
 }
 BOOST_FIXTURE_TEST_CASE_WITH_DECOR(DecryptWithECKeyV1, DecryptFixture,
                      * utf::depends_on("CDoc1ECKeySingle/EncryptWithECKeyV1")
                      * utf::description("Decrypting a file in CDoc1 format with with EC private key"))
 {
-    decrypt({checkDataFile(sources[0])}, checkTargetFile("ECKeyUsageV1.cdoc"), tmpDataPath.string(), fetchDataFile(ECPrivKeyFile));
+    decrypt({checkDataFile(sources[0])}, checkTargetFile("ECKeyUsageV1.cdoc"), tmpDataPath.string(), fetchDataFile(EC384PrivKeyFile));
 }
 BOOST_AUTO_TEST_SUITE_END()
 
@@ -636,13 +620,13 @@ BOOST_AUTO_TEST_SUITE(CDoc1ECKeyMulti)
 BOOST_FIXTURE_TEST_CASE_WITH_DECOR(EncryptWithECKeyV1Multi, EncryptFixture,
         * utf::description("Encrypting multiple files with EC key in CDoc1 format"))
 {
-    encryptV1({checkDataFile(sources[0]), checkDataFile(sources[1]), checkDataFile(sources[2])}, formTargetFile("ECKeyUsageV1Multi.cdoc"), fetchDataFile(ECCertFile));
+    encryptV1({checkDataFile(sources[0]), checkDataFile(sources[1]), checkDataFile(sources[2])}, formTargetFile("ECKeyUsageV1Multi.cdoc"), fetchDataFile(EC384CertFile));
 }
 BOOST_FIXTURE_TEST_CASE_WITH_DECOR(DecryptWithECKeyV1Multi, DecryptFixture,
                      * utf::depends_on("CDoc1ECKeyMulti/EncryptWithECKeyV1Multi")
                      * utf::description("Decrypting multiple files in CDoc1 format with with EC private key"))
 {
-    decrypt({checkDataFile(sources[0]), checkDataFile(sources[1]), checkDataFile(sources[2])}, checkTargetFile("ECKeyUsageV1Multi.cdoc"), tmpDataPath.string(), fetchDataFile(ECPrivKeyFile));
+    decrypt({checkDataFile(sources[0]), checkDataFile(sources[1]), checkDataFile(sources[2])}, checkTargetFile("ECKeyUsageV1Multi.cdoc"), tmpDataPath.string(), fetchDataFile(EC384PrivKeyFile));
 }
 BOOST_AUTO_TEST_SUITE_END()
